+++ client/src/extension.ts
@@ -5,8 +5,11 @@
 'use strict';
 
 import * as path from 'path';
+import * as fs from 'fs';
 
-import { window, workspace, Disposable, ExtensionContext, commands } from 'vscode';
+import { window, workspace, Disposable, ExtensionContext, commands, Uri } from 'vscode';
+import * as vscode from 'vscode';
+
 import { LanguageClient, LanguageClientOptions, SettingMonitor, ServerOptions, TransportKind } from 'vscode-languageclient';
 
 export function activate(context: ExtensionContext) {
@@ -27,5 +30,27 @@ export function activate(context: ExtensionContext) {
 
   const client = new LanguageClient('Api Elements', serverOptions, clientOptions);
 
+  context.subscriptions.push(commands.registerTextEditorCommand('apiElements.parserOutput', (editor) => {
+    const statusBarDisposable = window.setStatusBarMessage("Parsing current document...");
+
+    client.sendRequest({method: "parserOutput"}, editor.document.getText()).then((result) => {
+      const stringifiedResult = JSON.stringify(result, null, 2);
+      const uri = Uri.parse(`untitled:${workspace.rootPath ||context.extensionPath}/parseResult.json`);
+      workspace.openTextDocument(uri).then((textDocument) => {
+        const edit = new vscode.WorkspaceEdit();
+        edit.insert(uri, new vscode.Position(0,0), stringifiedResult);
+        workspace.applyEdit(edit).then(() => {
+          window.showTextDocument(textDocument, vscode.ViewColumn.One ,false).then(() => {
+            statusBarDisposable.dispose();
+          }, showError);
+        }, showError);
+      }, showError);
+    }, showError)
+  }));
+
   context.subscriptions.push(client.start());
+
+  function showError(err) {
+    window.showErrorMessage(err || err.message);
+  }
 }